+++ function_app.py
@@ -4,8 +4,8 @@ import os
 
 import azure.functions as func
 from azure.storage.filedatalake import FileSystemClient
-
 import config
+import moving_directories
 from publish_pipeline.generate_high_level_metadata.generate_changelog import (
     pipeline as generate_changelog_pipeline,
 )
@@ -172,39 +172,23 @@ def generate_discovery_metadata(req: func.HttpRequest) -> func.HttpResponse:
 @app.route(route="moving-folders", auth_level=func.AuthLevel.FUNCTION)
 def moving_folders(req: func.HttpRequest) -> func.HttpResponse:
     """Moves the directories along with the files in the Azure Database."""
-    overwrite_permitted = (
-        req.params["overwrite-permitted"]
-        if "overwrite-permitted" in req.params
-        else "true"
-    )
     file_system = FileSystemClient.from_connection_string(
         config.AZURE_STORAGE_CONNECTION_STRING,
         file_system_name="stage-1-container",
     )
-    new_dir_name: str = "AI-READI/metadata/test2/sub4"
-
     dir_name: str = "AI-READI/metadata/test2/sub5"
+    new_dir_name: str = "AI-READI/metadata/test2/sub4"
     source_path = file_system.get_directory_client(dir_name)
     destination_path = file_system.get_directory_client(new_dir_name)
 
-    if overwrite_permitted != "true" and overwrite_permitted != "false":
-        return func.HttpResponse(
-            "Only overwrite-permitted=true or overwrite-permitted=false accepted",
-            status_code=403,
+    moving_directories.moving_dirs(req, source_path, destination_path)
+    try:
+        source_path.rename_directory(
+            new_name=f"{source_path.file_system_name}/{new_dir_name}"
         )
-    if overwrite_permitted == "false" and destination_path.exists():
-        return func.HttpResponse(
-            "overwriting directories is not accepted", status_code=500
-        )
-    if destination_path.exists() and source_path.exists():
-        destination_path.delete_directory()
-        try:
-            source_path.rename_directory(
-                new_name=f"{source_path.file_system_name}/{new_dir_name}"
-            )
-            return func.HttpResponse("Success", status_code=200)
-        except Exception as e:
-            print(f"Exception: {e}")
+        return func.HttpResponse("Success", status_code=200)
+    except Exception as e:
+        print(f"Exception: {e}")
     return func.HttpResponse("Failed", status_code=500)
 
 
@@ -221,63 +205,21 @@ def copying_folders(req: func.HttpRequest) -> func.HttpResponse:
         file_system_name="stage-1-container",
     )
     dir_name: str = "AI-READI/metadata/test2/sub4"
-
     new_dir_name: str = "AI-READI/metadata/test2/sub5"
-
     directory_path = file_system.get_directory_client(dir_name)
     directory: str = directory_path.get_directory_properties().name
 
-    if overwrite_permitted != "true" and overwrite_permitted != "false":
-        return func.HttpResponse(
-            "Only overwrite-permitted=true or overwrite-permitted=false accepted",
-            status_code=403,
-        )
+    moving_directories.copying_dirs_beginning(req)
     try:
-        copy_directory(
+        moving_directories.copy_directory(
             file_system,
             directory,
             new_dir_name,
             True if overwrite_permitted.lower().strip() == "true" else False,
         )
         return func.HttpResponse("Success", status_code=200)
-
     except Exception as e:
         print(f"Exception: {e}")
-
         return func.HttpResponse("Failed", status_code=500)
 
 
-def copy_directory(
-    file_system: FileSystemClient,
-    source: str,
-    destination: str,
-    overwrite_permitted: bool,
-) -> None:
-    """Moving directories while implementing subsequent copies (recursion)"""
-    directory_client = file_system.get_directory_client(destination)
-    if destination.lower().startswith(source.lower()):
-        raise Exception("the destination is inside of the source")
-    if not directory_client.exists():
-        directory_client.create_directory()
-    else:
-        if not overwrite_permitted:
-            raise Exception("overwriting directories is not accepted")
-    for path in file_system.get_paths(source, recursive=False):
-        target = (
-            destination + "/" + os.path.basename(path.name.rstrip("/").rstrip("\\"))
-        )
-        if not path.is_directory:
-            source_file = file_system.get_file_client(path.name)
-
-            destination_file = file_system.get_file_client(target)
-
-            source_file_bytes = source_file.download_file().readall()
-
-            if not destination_file.exists():
-                destination_file.create_file()
-            destination_file.upload_data(source_file_bytes, overwrite=True)
-
-        else:
-            copy_directory(file_system, path.name, target, overwrite_permitted)
-            # check if dir is empty:
-            #     delete_directory
